backend/requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.9
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic[email]>=2.9.0
pydantic-settings>=2.6.0
supabase>=2.10.0
PyJWT>=2.9.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
openpyxl>=3.1.0
nameparser>=1.1.3
apify-client>=1.8.0
instaloader>=4.14.0
google-api-python-client>=2.130.0
